requirements.txt
asttokens==3.0.0
babel==2.17.0
backrefs==5.8
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.0
ghp-import==2.1.0
graphviz==0.20.3
griffe==1.7.3
idna==3.10
ipython==8.36.0
jedi==0.19.2
Jinja2==3.1.6
Markdown==3.8
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-autorefs==1.4.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.14
mkdocs-material-extensions==1.3.1
mkdocstrings==0.29.1
mkdocstrings-python==1.16.10
networkx==3.4.2
numpy==2.2.6
packaging==25.0
paginate==0.5.7
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pymdown-extensions==10.15
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyyaml_env_tag==1.1
requests==2.32.3
six==1.17.0
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.13.2
urllib3==2.4.0
watchdog==6.0.0
wcwidth==0.2.13
